Fix admin window
